Fix error using log writer in console commands
When using the log writer for logging errors in a console command the
following error is thrown:

In UserDataBag.php line 178: The "" value is not a valid IP address.

This comes from GeneralUtility::getIndpEnv('REMOTE_ADDR') which returns
an empty string as IP address on console. In this case the IP address of
localhost (127.0.0.1) is assigned instead.