Add PSR-3 variable interpolation to log writer
TYPO3 introduced an interpolate method in TYPO3 11 to implement
the PSR-3 variable interpolation in log messages. The interpolate
method must be called by the logger to replace placeholders in
the log message.

So if the method exists, which means we're in TYPO3 11, it is now
called in the logger.

Fixes #66